refactor: 코드 리뷰 반영 조건부 카드 스타일링으로 수정
