--- src/components/card/reviewCard/smallReview/SmallReview.tsx
@@ -1,38 +1,0 @@
-import * as styles from './smallReview.css';
-import CardInfo from '../cardInfo/cardInfo';
-
-interface SmallReviewProps {
-  reviewTitle: string;
-  reviewDate: string;
-  reviewName: string | null;
-  reviewLink: string;
-  blogImage: string | null;
-}
-
-const SmallReview = ({
-  reviewTitle,
-  reviewDate,
-  reviewName,
-  reviewLink,
-  blogImage,
-}: SmallReviewProps) => {
-  const handleButtonClick = () => {
-    window.location.href = reviewLink;
-  };
-
-  return (
-    <button className={styles.cardContainer} onClick={handleButtonClick}>
-      {blogImage ? (
-        <img className={styles.cardImage} src={blogImage} alt="thumbnail" />
-      ) : (
-        <img className={styles.cardImage} src="/default-image.png" alt="Default thumbnail" />
-      )}
-      <div className={styles.cardContent}>
-        <p className={styles.cardTitle}>{reviewTitle}</p>
-        <CardInfo reviewName={reviewName} reviewDate={reviewDate} />
-      </div>
-    </button>
-  );
-};
-
-export default SmallReview;
